requirements.txt
pandas
requests
requests-cache
openmeteo-requests
retry-requests
python-dotenv